Create a function for creating a Transaction from a CSV Line.
Refactor to create a separate method for converting a line of our CSV into a transaction.

It might be tempting to put this method inside the Transaction class itself (and then be able to call `Transaction.FromCsvLine(line)` or similar.

I don't think that's necessarily wrong... but I'd argue that our 'transaction' class shouldn't need to worry about where the data came from or what format the CSV file is in etc. - that's the job of the File Reader, so is where I've chosen to put this method.