Use Linq in place of foreach loops!
Ok... so this one is my favourite, but perhaps a little 'new dev' unfriendly!

Linq is really awesome! but definitely takes some getting used to. Initially the learning curve is high, but once you've cracked it I promise its worth it.

If you are comfortable with Linq, the code produced is
- MUCH more concise.
- easier to understand
- less likely to contain bugs.

Take the sum as an example. Conceptually its a little trickier to figure out how its working (with a foreach loop it works the same way you'd do it in your head)

but... then if you've got that sorted then the Linq approach is awesome!
- 1 line of code in place of 6
- and super clear - this line takes the transactions, and then sums their Amounts.

Clear - Expressive - Concise! Awesome!

If you want to learn more about Linq, then this is a good resource:
https://visualstudiomagazine.com/articles/2010/03/01/using-linq-to-express-intent.aspx


The second change that happened at the same time was that we switched from using Lists or Sets to using IEnumerable.

First up - what is an IEnumerable? In short, its anything that I can loop through! So it includes things like lists, sets dictionaries etc, but is a much more general thing. Basically if I could throw it in a foreach loop then its an IEnumerable! :) 

This is a subtle change though, and one that's always tricky to describe! (happy to chat in person if you like!)

The idea is that we should be as general as we can be with our types.
ie - if we have a list of names, but all we ever do is loop over that names, then do we care that it was actually a list? Would anything need to change if we had passed in a set instead?
If the answer is no? then why did we specify that you should pass in a list? If our method doesn't actually care, then why are we being needlessly specific.

We could, instead change that input from being a list to being an IEnumerable. Now we can pass in anything we like (as long as we can loop over it). This means that we can re-use our method in more ways - happy days!